Add an "Extending this mechanism" section
Rick and I decided to provide some extensibility to enable future user identities to be represented via this mechanism, but it turns out that I don't have specific answers on a number of questions we have about how to proceed, so I can't write this section yet. But let's add a place-holder.